Refactor Accessibility components (#11)
Moving `useAnnouncement` within a sub-component called Accessibility in order to avoid having the parent DndContext component re-render whenever we append and remove announcements to be read from the LiveRegion.

In an ideal world, announcements would be done in a more declarative fashion, but currently there is no way to distinguish declaratively between a drag end and drag cancel event, so unfortunately we need to `useImperativeHandle` to call the `announce` method imperatively. This is a bit of a code smell, but we can keep this approach for now until we can refactor `useDrag` to declaratively be able to distinguish between an end and cancel event.