Fix `getEventListenerTarget` when target element is not an instance of `HTMLElement` (#335)
For example, when the target element is an SVGElement, the event listener target returned was the owner document instead.

This causes issues with touch events when the `target` is an SVGElement. If the `event.target` element is removed from the document events will still be targeted at it, and hence won't always bubble up to the window or document anymore. If there is any risk of an element being removed while it is being dragged, the best practice is to attach the event listeners directly to the target.