Add key to DraggableNode and DroppableContainer interfaces
This prevents potential race conditions in the mount and cleanup effects of `useDraggable` and `useDroppable`. It's possible for the clean-up effect to run after another React component using `useDraggable` or `useDroppable` mounts, which causes the newly mounted element to accidentally be un-registered.

Adding a unique key that is tied to the component instance where the `useDraggable` or `useDroppable` hook is invoked fixes these types of race conditions.