stories: fix edge case in multiple containers collision detection
When a draggable item moves to a new container, the layout may shift and the `overId` may become `null`. We manually set the cached `lastOverId` to the id of the draggable item that was moved to the new container, otherwise the previous `overId` will be returned which can cause items to incorrectly shift positions.