Observe resizes of droppable containers while dragging (#561)
* Allow partial recomputation of droppable rects

* Add resize observer to droppable elements while dragging

While sorting, it's possible for the size of droppable elements to change. We add a resize observer to keep track of size changes and schedule recomputation of the bounding rectangles of the droppable elements when resize events occur.

The `resizeObserverConfig` param can be used to disable or configure the resize observer.