Fix drawer story
The original implementation was relying on a quirk of how the container rect used to be measured without considering transforms. The new measuring strategy is no longer transform agnostic, therefore we cannot rely on the `top` coordinates of the container node rect as they include transforms.

In the future, it may be possible to open up configuring how the container node rect is measured to consumers if the need arises.